Create: Crafted Beginning v0.0.9 更新日志 ====================
====================
新增内容
■ 注气室滤网
可以安装在注气室底部，从而让注气室实现鼓风机类似的批量加工行为。
需要先用鼓风机的加工方式处理后才能记录对应的加工类型。

修改内容
■ 气密板
修改序列组装配方，调换了第一个材料（铁板）和第二个材料（粘液球或者强力胶）的顺序。

■ 气密壳装管道
现在有更清晰的可视化显示。

■ 旋风人蓄风室
现在实时根据恢复的饥饿值、饱和度和提供的药水效果计算食物可以提供的时长。
现在自动支持其他模组添加的食物。
重新设计蓄风配方，现在需要额外指定action字段为charge、clear_ill或者cycle_creative。
略微调整了时长算法，并允许在配方中返还一个物品。

■ 气密锻造机
现在可以执行与辊压机相同的辊压配方，可通过配置选项开关。

■ 注气室
现在允许进行批量注气，单次可加工最多一组物品。

■ 气体工厂仪表
修改合成配方，现在材料中还额外需要紫晶体板。

■ 气球
重新调整工具栏显示，不再用耐久条显示气体含量与种类。

■ 终末匿音机
现在声音事件的来源或振动监听者之一位于工作范围内就会被阻止。
现在兼容动态结构，不仅可以在动态结构上正常工作，而且可以使用动态结构控制器开关。
现在兼容机械动力：航空学，在物理结构上也会正常工作。

■ 创造物品栏
现在根据"气密系列"、"光应力系列"、"终末系列"、"装饰"和"气罐"进行分类。

====================
Bug修复
■ 修复在安装iris时黑曜石砖块N、黑曜石砖块S、哭泣黑曜石砖块N和哭泣黑曜石砖块S丢失材质的问题。
■ 修复特斯拉涡轮机和气密引擎的护目镜提示框里在其他字体或英文语言环境下不对齐的问题。
■ 修复注气序列组装第一个材料显示出错的问题。
■ 修复注气室动画卡顿的问题。
■ 修复气密引擎动画卡顿的问题。
■ 修复在未安装JEI时右键打开仓库管理员的GUI会报错闪退的问题。
■ 修复气体过滤器的最大堆叠数被设置为1的问题。
■ 修复在安装流体包裹且进行JEI转单时会丢失气体材料的问题。
■ 修复气球显示气体含量时冒号的颜色也被渲染成金色的问题。
■ 修复在重新进入存档后气密引擎产生的应力量异常的问题。

+++ src/main/java/net/ty/createcraftedbeginning/content/airtights/airtightcannon/AirtightCannonUtils.java
@@ -78,9 +78,11 @@ public final class AirtightCannonUtils {
         if (owner == null) {
             return false;
         }
+
         if (target == owner || owner.isAlliedTo(target) || target.isAlliedTo(owner)) {
             return true;
         }
+
         if (!(target instanceof TamableAnimal animal)) {
             return false;
         }
@@ -119,7 +121,6 @@ public final class AirtightCannonUtils {
         if (!CanisterContainerSuppliers.isAnyContainerAvailable(player) || gasContent.isEmpty() || timeCharged < MIN_USE_TIME) {
             return OptionalDouble.empty();
         }
-
         return OptionalDouble.of(Mth.clamp((double) timeCharged / getEfficientUseTime(cannon), 0, 2));
     }
 
@@ -132,13 +133,12 @@ public final class AirtightCannonUtils {
         if (!cannon.is(CCBItems.AIRTIGHT_CANNON)) {
             return 0;
         }
-
         return cannon.getTagEnchantments().entrySet().stream().filter(entry -> entry.getKey().is(enchantment)).findFirst().map(Entry::getValue).orElse(0);
     }
 
     public static double getRawGasConsumption(ItemStack cannon, float chargedRatio) {
         int multiShotMultiplier = 2 * getEnchantmentLevel(cannon, Enchantments.MULTISHOT) + 1;
-        double ratio = chargedRatio >= 1 ? Mth.square(chargedRatio) : Mth.sqrt(chargedRatio);
+        float ratio = chargedRatio >= 1 ? Mth.square(chargedRatio) : Mth.sqrt(chargedRatio);
         int gasConsumption = CCBConfig.server().equipments.perShotConsumption.get();
         return gasConsumption * (double) multiShotMultiplier * ratio;
     }
